+++ zlabel/widgets/ui/dialog_settings.py
@@ -280,6 +280,12 @@ class Ui_DialogSettings(object):
 
         self.grid_performance.addWidget(self.spin_pyramid_levels, 0, 5, 1, 1)
 
+        self.ckbox_mipmap = QCheckBox(self.gbox_performance)
+        self.ckbox_mipmap.setObjectName(u"ckbox_mipmap")
+        self.ckbox_mipmap.setChecked(True)
+
+        self.grid_performance.addWidget(self.ckbox_mipmap, 4, 0, 1, 2)
+
 
         self.gridLayout.addWidget(self.gbox_performance, 3, 0, 1, 1)
 
@@ -838,6 +844,7 @@ class Ui_DialogSettings(object):
         self.label_tl_small_side.setText(QCoreApplication.translate("DialogSettings", u"Timeline small side:", None))
         self.spin_tl_small_side.setSuffix(QCoreApplication.translate("DialogSettings", u"px", None))
         self.label_pyramid_levels.setText(QCoreApplication.translate("DialogSettings", u"Pyramid levels:", None))
+        self.ckbox_mipmap.setText(QCoreApplication.translate("DialogSettings", u"Enable Mipmap Anti-aliasing", None))
         self.gbox_appearance.setTitle(QCoreApplication.translate("DialogSettings", u"Appearance", None))
         self.label.setText(QCoreApplication.translate("DialogSettings", u"Fill alpha:", None))
         self.label_default_color.setText(QCoreApplication.translate("DialogSettings", u"Default color:", None))